+++ src/ts/Ads/Parsers/AuctionResponseParser.spec.ts
@@ -15,6 +15,7 @@ describe('AuctionResponseParser', () => {
             'premium': Placement('premium'),
             'video': Placement('video'),
             'rewardedVideoZone': Placement('rewardedVideoZone'),
+            'outOfBoundsStartPlacement': Placement('outOfBoundsStartPlacement'),
             'mraid': Placement('mraid')
         };
     });
@@ -53,14 +54,17 @@ describe('AuctionResponseParser', () => {
             });
 
             it('should contain the correct amount of AuctionPlacements', () => {
-                expect(parsedAuctionResponse.auctionResponses[0].getPlacements().length).toEqual(3);
+                expect(parsedAuctionResponse.auctionResponses[0].getPlacements().length).toEqual(4);
             });
 
-            it('should contain the correct placementId', () => {
+            it('should contain the correct placementIds', () => {
                 expect(parsedAuctionResponse.auctionResponses[0].getPlacements()[0].getPlacementId()).toEqual('premium');
+                expect(parsedAuctionResponse.auctionResponses[0].getPlacements()[1].getPlacementId()).toEqual('video');
+                expect(parsedAuctionResponse.auctionResponses[0].getPlacements()[2].getPlacementId()).toEqual('rewardedVideoZone');
+                expect(parsedAuctionResponse.auctionResponses[0].getPlacements()[3].getPlacementId()).toEqual('outOfBoundsStartPlacement');
             });
 
-            it('should contain the correct trackingUrls', () => {
+            it('should contain the correct trackingUrls for premium', () => {
                 const expectedTrackingUrls = {
                     start: [
                         'https://tracking.stg.mz.internal.unity3d.com/operative/placement=premiumAuction?eventType=start',
@@ -71,6 +75,36 @@ describe('AuctionResponseParser', () => {
                     ]
                 };
                 expect(parsedAuctionResponse.auctionResponses[0].getPlacements()[0].getTrackingUrls()).toStrictEqual(expectedTrackingUrls);
+            });
+
+            it('should contain the correct trackingUrls for video', () => {
+                const expectedTrackingUrls = {
+                    impression: [
+                        'https://tracking.stg.mz.internal.unity3d.com/operative/placement=video?eventType={{eventType}}'
+                    ]
+                };
+                expect(parsedAuctionResponse.auctionResponses[0].getPlacements()[1].getTrackingUrls()).toStrictEqual(expectedTrackingUrls);
+            });
+
+            it('should contain the correct trackingUrls for rewardedVideoZone', () => {
+                const expectedTrackingUrls = {
+                    unsupportedEvent: [
+                        'https://ads.brand.postback.unity3d.com/impression/placement=rewardedVideoZone?data=rewardedVideoZoneDataBlob?secret=fifthQuartile'
+                    ]
+                };
+                expect(parsedAuctionResponse.auctionResponses[0].getPlacements()[2].getTrackingUrls()).toStrictEqual(expectedTrackingUrls);
+            });
+
+            it('should contain the correct trackingUrls for outOfBoundsStartPlacement with two out of bounds entries', () => {
+                const expectedTrackingUrls = {
+                    start: [
+                        'https://ads.brand.postback.unity3d.com/impression/placement=oobPlacement?data=outOfBoundsData?secret=scott'
+                    ],
+                    complete: [
+                        'https://tracking.stg.mz.internal.unity3d.com/operative/placement=oobPlacement?eventType=complete'
+                    ]
+                };
+                expect(parsedAuctionResponse.auctionResponses[0].getPlacements()[3].getTrackingUrls()).toStrictEqual(expectedTrackingUrls);
             });
         });
     });