+++ src/ts/Ads/Views/OpenMeasurement/AdmobOpenMeasurementController.ts
@@ -160,10 +160,16 @@ export class AdmobOpenMeasurementController extends OpenMeasurementController {
     }
 
     public start(duration: number) {
-        if (!this._omMetricFirstCheck && this._platform === Platform.ANDROID) {
-            SDKMetrics.reportMetricEvent(AdmobMetric.AdmobOMStartFirst);
-            this._omMetricFirstCheck = true;
-        }
+
+        // timeout used to experiment how race condition can be massaged
+        // until admob creatives are fixed
+        setTimeout(() => {
+            if (!this._omMetricFirstCheck && this._platform === Platform.ANDROID) {
+                SDKMetrics.reportMetricEvent(AdmobMetric.AdmobOMStartFirst);
+                this._omMetricFirstCheck = true;
+            }
+        }, 50);
+
         super.start(duration);
     }