+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Pending
 
+* [Auction V6] Add Auction Response Parser [#14033](https://github.com/Applifier/unity-ads-webview/pull/14033)
+* [Auction V6] Use Auction Parser in LegacyCampaignManager [#14034](https://github.com/Applifier/unity-ads-webview/pull/14034)
+* [Load V5] Set Retry to 0 for Requests [#14038](https://github.com/Applifier/unity-ads-webview/pull/14038)
+* [Load V5] Mocks for PerPlacementLoadManagerV5 tests [#14039](https://github.com/Applifier/unity-ads-webview/pull/14039)
+* [Load V5] Increment Ad Ordinal [#14036](https://github.com/Applifier/unity-ads-webview/pull/14036)
+* Open Measurement - Add Timeout on Admob start metric [#14040](https://github.com/Applifier/unity-ads-webview/pull/14040)
+* Update AndroidDeviceInfo DTO [#14001](https://github.com/Applifier/unity-ads-webview/pull/14001)
+
+# Thu Mar 26 23:11:54 UTC 2020
+
 * [Auction V6] Add Mocks [#14032](https://github.com/Applifier/unity-ads-webview/pull/14032)
 * [Load V5] Ad Request Manager [#13881](https://github.com/Applifier/unity-ads-webview/pull/13881)
 * Add whitelist for Apple PMP creative [#13954](https://github.com/Applifier/unity-ads-webview/pull/13954)
@@ -8,6 +18,8 @@
 * Open Measurement - Add Metrics for Admob OM Start and Loaded [#13998](https://github.com/Applifier/unity-ads-webview/pull/13998)
 * Sample SDKMetrics at 50% [#14031](https://github.com/Applifier/unity-ads-webview/pull/14031)
 * Remove SDKDetectionInfo From Request Body [#14027](https://github.com/Applifier/unity-ads-webview/pull/14027)
+* Add CacheMode Allowed Reverse ABTest for Whitelisted Games [#13992](https://github.com/Applifier/unity-ads-webview/pull/13992)
+* Remove Sending "Negative Value" Metrics to SDKMetrics [#14037](https://github.com/Applifier/unity-ads-webview/pull/14037)
 
 # Wed Mar 25 21:40:32 UTC 2020